[#1835] Move equipped and atunement to seperate template
