+++ module/data/item/_module.mjs
@@ -25,6 +25,7 @@ export {
 };
 export {default as ActionTemplate} from "./templates/action.mjs";
 export {default as ActivatedEffectTemplate} from "./templates/activated-effect.mjs";
+export {default as EquippableItemTemplate} from "./templates/equippable-item.mjs";
 export {default as ItemDescriptionTemplate} from "./templates/item-description.mjs";
 export {default as MountableTemplate} from "./templates/mountable.mjs";
 export {default as PhysicalItemTemplate} from "./templates/physical-item.mjs";
